+++ CHANGELOG.md
@@ -9,6 +9,7 @@
   `pgc_model:column/3` for the same use case.
 - Add default values to models.
 - Add support for enum values.
+- Add support for domain values.
 # Bugs
 - Fix handling of unknown `gen_server` calls.
 - Delete the model registry ETS table during shutdown.